main.py


        
import os,sys,re,json,time, numpy, pandas,csv, yaml
import tweepy
import logging
import traceback
from datetime import datetime,timezone
from dateutil.tz import tzlocal
from dateutil.parser import parse
def main():
    import src.run
    src.run.main()

main()