requirements.txt
pyinstaller
pyinstaller-hooks-contrib
tweepy
numpy==1.19.3
pandas
pyyaml